+++ src/parser/inline/deletion.test.ts
@@ -12,19 +12,20 @@ describe('Unit: parser/inline/deletion', () => {
       assert.deepStrictEqual(inspect(parser('~~')), undefined);
       assert.deepStrictEqual(inspect(parser('~~a')), [['~~', 'a'], '']);
       assert.deepStrictEqual(inspect(parser('~~a~')), [['~~', 'a', '~'], '']);
-      assert.deepStrictEqual(inspect(parser('~~ ~~')), undefined);
-      assert.deepStrictEqual(inspect(parser('~~ a~~')), undefined);
-      assert.deepStrictEqual(inspect(parser('~~ a ~~')), undefined);
-      assert.deepStrictEqual(inspect(parser('~~\na~~')), undefined);
-      assert.deepStrictEqual(inspect(parser('~~\\\na~~')), undefined);
-      assert.deepStrictEqual(inspect(parser('~~<wbr>a~~')), undefined);
-      assert.deepStrictEqual(inspect(parser('~~<# a #>b~~')), undefined);
     });
 
     it('basic', () => {
       assert.deepStrictEqual(inspect(parser('~~a~~')), [['<del>a</del>'], '']);
       assert.deepStrictEqual(inspect(parser('~~ab~~')), [['<del>ab</del>'], '']);
       assert.deepStrictEqual(inspect(parser('~~a ~~')), [['<del>a </del>'], '']);
+      assert.deepStrictEqual(inspect(parser('~~ ~~')), [['<del> </del>'], '']);
+      assert.deepStrictEqual(inspect(parser('~~ a~~')), [['<del> a</del>'], '']);
+      assert.deepStrictEqual(inspect(parser('~~ a ~~')), [['<del> a </del>'], '']);
+      assert.deepStrictEqual(inspect(parser('~~  a  ~~')), [['<del>  a  </del>'], '']);
+      assert.deepStrictEqual(inspect(parser('~~\na~~')), [['<del><br>a</del>'], '']);
+      assert.deepStrictEqual(inspect(parser('~~\\\na~~')), [['<del><span class="linebreak"> </span>a</del>'], '']);
+      assert.deepStrictEqual(inspect(parser('~~<wbr>a~~')), [['<del><wbr>a</del>'], '']);
+      assert.deepStrictEqual(inspect(parser('~~<# a #>b~~')), [['<del><sup class="comment" title="a"></sup>b</del>'], '']);
       assert.deepStrictEqual(inspect(parser('~~a\nb~~')), [['<del>a<br>b</del>'], '']);
       assert.deepStrictEqual(inspect(parser('~~a\\\nb~~')), [['<del>a<span class="linebreak"> </span>b</del>'], '']);
       assert.deepStrictEqual(inspect(parser('~~\\~~~')), [['<del>~</del>'], '']);