+++ src/parser/inline/deletion.ts
@@ -1,16 +1,13 @@
 import { DeletionParser, inline } from '../inline';
 import { union, some, creator, surround, lazy } from '../../combinator';
-import { startTight, isEndTight, trimEnd, defrag } from '../util';
+import { defrag } from '../util';
 import { str } from '../source';
 import { html } from 'typed-dom';
 import { unshift } from 'spica/array';
 
 export const deletion: DeletionParser = lazy(() => creator(surround(
   str('~~'),
-  startTight(union([some(inline, '~~')])),
+  union([some(inline, '~~')]),
   str('~~'), false,
-  ([as, bs, cs], rest) =>
-    isEndTight(bs)
-      ? [[html('del', defrag(trimEnd(bs)))], rest]
-      : [unshift(as, bs), cs[0] + rest],
+  ([, bs], rest) => [[html('del', defrag(bs))], rest],
   ([as, bs], rest) => [unshift(as, bs), rest])));